+++ python/cugraph/__init__.py
@@ -26,8 +26,8 @@ from cugraph.cores import core_number, k_core
 from cugraph.components import weakly_connected_components, strongly_connected_components
 from cugraph.ktruss import ktruss_max, ktruss_subgraph
 from cugraph.link_analysis import pagerank
+from cugraph.structure import Graph, DiGraph, from_cudf_edgelist, renumber, symmetrize, symmetrize_df
 from cugraph.link_prediction import jaccard, overlap, jaccard_w, overlap_w
-from cugraph.structure import Graph, from_cudf_edgelist, renumber, symmetrize, symmetrize_df
 from cugraph.traversal import bfs, sssp, filter_unreachable
 # from cugraph.utilities import grmat_gen
 from cugraph.utilities import device_of_gpu_pointer