+++ python/cugraph/community/spectral_clustering_wrapper.pyx
@@ -18,6 +18,7 @@
 
 from cugraph.community.c_spectral_clustering cimport *
 from cugraph.structure.c_graph cimport *
+from cugraph.structure import graph_wrapper
 from cugraph.utilities.column_utils cimport *
 from libcpp cimport bool
 from libc.stdint cimport uintptr_t
@@ -30,7 +31,7 @@ import rmm
 import numpy as np
 
 
-def spectralBalancedCutClustering(graph_ptr,
+def spectralBalancedCutClustering(input_graph,
                                     num_clusters,
                                     num_eigen_vects=2,
                                     evs_tolerance=.00001,
@@ -40,13 +41,23 @@ def spectralBalancedCutClustering(graph_ptr,
     """
     Call balancedCutClustering_nvgraph
     """
-
-    cdef uintptr_t graph = graph_ptr
+    cdef uintptr_t graph = graph_wrapper.allocate_cpp_graph()
     cdef Graph * g = <Graph*> graph
 
-    # Ensure that the graph has CSR adjacency list
-    add_adj_list(g)
-    
+    if input_graph.adjlist:
+        [offsets, indices] = graph_wrapper.datatype_cast([input_graph.adjlist.offsets, input_graph.adjlist.indices], [np.int32])
+        [weights] = graph_wrapper.datatype_cast([input_graph.adjlist.weights], [np.float32, np.float64])
+        graph_wrapper.add_adj_list(graph, offsets, indices, weights)
+    else:
+        [src, dst] = graph_wrapper.datatype_cast([input_graph.edgelist.edgelist_df['src'], input_graph.edgelist.edgelist_df['dst']], [np.int32])
+        if input_graph.edgelist.weights:
+            [weights] = graph_wrapper.datatype_cast([input_graph.edgelist.edgelist_df['weights']], [np.float32, np.float64])
+            graph_wrapper.add_edge_list(graph, src, dst, weights)
+        else:
+            graph_wrapper.add_edge_list(graph, src, dst)
+        add_adj_list(g)
+        offsets, indices, values = graph_wrapper.get_adj_list(graph)
+        input_graph.adjlist = input_graph.AdjList(offsets, indices, values)
 
     # we should add get_number_of_vertices() to Graph (and this should be
     # used instead of g.adjList.offsets.size - 1)
@@ -73,9 +84,12 @@ def spectralBalancedCutClustering(graph_ptr,
                                             &c_cluster_col)
     
 
+    if input_graph.renumbered:
+        df['vertex'] = input_graph.edgelist.renumber_map[df['vertex']]
+
     return df
 
-def spectralModularityMaximizationClustering(graph_ptr,
+def spectralModularityMaximizationClustering(input_graph,
                                                num_clusters,
                                                num_eigen_vects=2,
                                                evs_tolerance=.00001,
@@ -85,13 +99,19 @@ def spectralModularityMaximizationClustering(graph_ptr,
     """
     Call spectralModularityMaximization_nvgraph
     """
-
-    cdef uintptr_t graph = graph_ptr
+    cdef uintptr_t graph = graph_wrapper.allocate_cpp_graph()
     cdef Graph * g = <Graph*> graph
 
-    # Ensure that the graph has CSR adjacency list
-    add_adj_list(g)
-    
+    if input_graph.adjlist:
+        graph_wrapper.add_adj_list(graph, input_graph.adjlist.offsets, input_graph.adjlist.indices, input_graph.adjlist.weights)
+    else:
+        if input_graph.edgelist.weights:
+            graph_wrapper.add_edge_list(graph, input_graph.edgelist.edgelist_df['src'], input_graph.edgelist.edgelist_df['dst'], input_graph.edgelist.edgelist_df['weights'])
+        else:
+            graph_wrapper.add_edge_list(graph, input_graph.edgelist.edgelist_df['src'], input_graph.edgelist.edgelist_df['dst'])
+        add_adj_list(g)
+        offsets, indices, values = graph_wrapper.get_adj_list(graph)
+        input_graph.adjlist = input_graph.AdjList(offsets, indices, values)
 
     # we should add get_number_of_vertices() to Graph (and this should be
     # used instead of g.adjList.offsets.size - 1)
@@ -118,18 +138,28 @@ def spectralModularityMaximizationClustering(graph_ptr,
                                                      &c_cluster_col)
     
 
+    if input_graph.renumbered:
+        df['vertex'] = input_graph.edgelist.renumber_map[df['vertex']]
+
     return df
 
-def analyzeClustering_modularity(graph_ptr, n_clusters, clustering):
+def analyzeClustering_modularity(input_graph, n_clusters, clustering):
     """
     Call analyzeClustering_modularity_nvgraph
     """
-    cdef uintptr_t graph = graph_ptr
+    cdef uintptr_t graph = graph_wrapper.allocate_cpp_graph()
     cdef Graph * g = <Graph*> graph
 
-    # Ensure that the graph has CSR adjacency list
-    add_adj_list(g)
-    
+    if input_graph.adjlist:
+        graph_wrapper.add_adj_list(graph, input_graph.adjlist.offsets, input_graph.adjlist.indices, input_graph.adjlist.weights)
+    else:
+        if input_graph.edgelist.weights:
+            graph_wrapper.add_edge_list(graph, input_graph.edgelist.edgelist_df['src'], input_graph.edgelist.edgelist_df['dst'], input_graph.edgelist.edgelist_df['weights'])
+        else:
+            graph_wrapper.add_edge_list(graph, input_graph.edgelist.edgelist_df['src'], input_graph.edgelist.edgelist_df['dst'])
+        add_adj_list(g)
+        offsets, indices, values = graph_wrapper.get_adj_list(graph)
+        input_graph.adjlist = input_graph.AdjList(offsets, indices, values)
 
     cdef gdf_column c_clustering_col = get_gdf_column_view(clustering)
     cdef float score
@@ -137,16 +167,23 @@ def analyzeClustering_modularity(graph_ptr, n_clusters, clustering):
     
     return score
 
-def analyzeClustering_edge_cut(graph_ptr, n_clusters, clustering):
+def analyzeClustering_edge_cut(input_graph, n_clusters, clustering):
     """
     Call analyzeClustering_edge_cut_nvgraph
     """
-    cdef uintptr_t graph = graph_ptr
+    cdef uintptr_t graph = graph_wrapper.allocate_cpp_graph()
     cdef Graph * g = <Graph*> graph
 
-    # Ensure that the graph has CSR adjacency list
-    add_adj_list(g)
-    
+    if input_graph.adjlist:
+        graph_wrapper.add_adj_list(graph, input_graph.adjlist.offsets, input_graph.adjlist.indices, input_graph.adjlist.weights)
+    else:
+        if input_graph.edgelist.weights:
+            graph_wrapper.add_edge_list(graph, input_graph.edgelist.edgelist_df['src'], input_graph.edgelist.edgelist_df['dst'], input_graph.edgelist.edgelist_df['weights'])
+        else:
+            graph_wrapper.add_edge_list(graph, input_graph.edgelist.edgelist_df['src'], input_graph.edgelist.edgelist_df['dst'])
+        add_adj_list(g)
+        offsets, indices, values = graph_wrapper.get_adj_list(graph)
+        input_graph.adjlist = input_graph.AdjList(offsets, indices, values)
 
     cdef gdf_column c_clustering_col = get_gdf_column_view(clustering)
     cdef float score
@@ -154,16 +191,23 @@ def analyzeClustering_edge_cut(graph_ptr, n_clusters, clustering):
     
     return score
 
-def analyzeClustering_ratio_cut(graph_ptr, n_clusters, clustering):
+def analyzeClustering_ratio_cut(input_graph, n_clusters, clustering):
     """
     Call analyzeClustering_ratio_cut_nvgraph
     """
-    cdef uintptr_t graph = graph_ptr
+    cdef uintptr_t graph = graph_wrapper.allocate_cpp_graph()
     cdef Graph * g = <Graph*> graph
 
-    # Ensure that the graph has CSR adjacency list
-    add_adj_list(g)
-    
+    if input_graph.adjlist:
+        graph_wrapper.add_adj_list(graph, input_graph.adjlist.offsets, input_graph.adjlist.indices, input_graph.adjlist.weights)
+    else:
+        if input_graph.edgelist.weights:
+            graph_wrapper.add_edge_list(graph, input_graph.edgelist.edgelist_df['src'], input_graph.edgelist.edgelist_df['dst'], input_graph.edgelist.edgelist_df['weights'])
+        else:
+            graph_wrapper.add_edge_list(graph, input_graph.edgelist.edgelist_df['src'], input_graph.edgelist.edgelist_df['dst'])
+        add_adj_list(g)
+        offsets, indices, values = graph_wrapper.get_adj_list(graph)
+        input_graph.adjlist = input_graph.AdjList(offsets, indices, values)
 
     cdef gdf_column c_clustering_col = get_gdf_column_view(clustering)
     cdef float score