+++ python/cugraph/link_analysis/pagerank_wrapper.pyx
@@ -22,23 +22,35 @@ from cugraph.utilities.column_utils cimport *
 from libcpp cimport bool
 from libc.stdint cimport uintptr_t
 from libc.stdlib cimport calloc, malloc, free
-
+from cugraph.structure import graph_wrapper
 import cudf
 import cudf._lib as libcudf
 import rmm
 import numpy as np
 
 
-def pagerank(graph_ptr,alpha=0.85, personalization=None, max_iter=100, tol=1.0e-5, nstart=None):
+def pagerank(input_graph, alpha=0.85, personalization=None, max_iter=100, tol=1.0e-5, nstart=None):
     """
     Call pagerank
     """
 
-    cdef uintptr_t graph = graph_ptr
-    cdef Graph* g = <Graph*>graph
+    cdef uintptr_t graph = graph_wrapper.allocate_cpp_graph()
+    cdef Graph * g = <Graph*> graph
 
-    add_transposed_adj_list(g)
-    
+    if input_graph.transposedadjlist:
+        [offsets, indices] = graph_wrapper.datatype_cast([input_graph.transposedadjlist.offsets, input_graph.transposedadjlist.indices], [np.int32])
+        [weights] = graph_wrapper.datatype_cast([input_graph.transposedadjlist.weights], [np.float32, np.float64])
+        graph_wrapper.add_transposed_adj_list(graph, offsets, indices, weights)
+    else:
+        [src, dst] = graph_wrapper.datatype_cast([input_graph.edgelist.edgelist_df['src'], input_graph.edgelist.edgelist_df['dst']], [np.int32])
+        if input_graph.edgelist.weights:
+            [weights] = graph_wrapper.datatype_cast([input_graph.edgelist.edgelist_df['weights']], [np.float32, np.float64])
+            graph_wrapper.add_edge_list(graph, src, dst, weights)    
+        else:
+            graph_wrapper.add_edge_list(graph, src, dst)
+        add_transposed_adj_list(g)
+        offsets, indices, values = graph_wrapper.get_transposed_adj_list(graph)
+        input_graph.transposedadjlist = input_graph.transposedAdjList(offsets, indices, values)
 
     # we should add get_number_of_vertices() to Graph (and this should be
     # used instead of g.transposedAdjList.offsets.size - 1)
@@ -60,7 +72,6 @@ def pagerank(graph_ptr,alpha=0.85, personalization=None, max_iter=100, tol=1.0e-
 
     g.transposedAdjList.get_vertex_identifiers(&c_identifier_col)
     
-
     if personalization is None:
         c_pagerank.pagerank(g, &c_pagerank_col, <gdf_column*> NULL, <gdf_column*> NULL,
                 <float> alpha, <float> tol, <int> max_iter, has_guess)
@@ -70,6 +81,6 @@ def pagerank(graph_ptr,alpha=0.85, personalization=None, max_iter=100, tol=1.0e-
         c_pagerank.pagerank(g, &c_pagerank_col, &c_pers_vtx, &c_pers_val,
                 <float> alpha, <float> tol, <int> max_iter, has_guess)
 
-    
-
+    if input_graph.renumbered:
+        df['vertex'] = input_graph.edgelist.renumber_map[df['vertex']]
     return df