+++ python/cugraph/link_prediction/woverlap_wrapper.pyx
@@ -18,6 +18,7 @@
 
 from cugraph.link_prediction.c_overlap cimport *
 from cugraph.structure.c_graph cimport *
+from cugraph.structure import graph_wrapper
 from cugraph.utilities.column_utils cimport *
 from cudf._lib.cudf cimport np_dtype_from_gdf_column
 from libc.stdint cimport uintptr_t
@@ -30,16 +31,27 @@ import numpy as np
 from cython cimport floating
 
 
-def overlap_w(graph_ptr, weights, first=None, second=None):
+def overlap_w(input_graph, weights, first=None, second=None):
     """
     Call overlap_list
     """
-
-    cdef uintptr_t graph = graph_ptr
+    cdef uintptr_t graph = graph_wrapper.allocate_cpp_graph()
     cdef Graph * g = <Graph*> graph
 
-    add_adj_list(g)
-    
+    if input_graph.adjlist:
+        [offsets, indices] = graph_wrapper.datatype_cast([input_graph.adjlist.offsets, input_graph.adjlist.indices], [np.int32])
+        [weights] = graph_wrapper.datatype_cast([input_graph.adjlist.weights], [np.float32, np.float64])
+        graph_wrapper.add_adj_list(graph, offsets, indices, weights)
+    else:
+        [src, dst] = graph_wrapper.datatype_cast([input_graph.edgelist.edgelist_df['src'], input_graph.edgelist.edgelist_df['dst']], [np.int32])
+        if input_graph.edgelist.weights:
+            [weights] = graph_wrapper.datatype_cast([input_graph.edgelist.edgelist_df['weights']], [np.float32, np.float64])
+            graph_wrapper.add_edge_list(graph, src, dst, weights)
+        else:
+            graph_wrapper.add_edge_list(graph, src, dst)
+        add_adj_list(g)
+        offsets, indices, values = graph_wrapper.get_adj_list(graph)
+        input_graph.adjlist = input_graph.AdjList(offsets, indices, values)
 
     cdef gdf_column c_result_col
     cdef gdf_column c_weight_col
@@ -88,6 +100,11 @@ def overlap_w(graph_ptr, weights, first=None, second=None):
         g.adjList.get_source_indices(&c_index_col);
         
         df['destination'] = cudf.Series(dest_data)
+
+        if input_graph.renumbered:
+            df['source'] = input_graph.edgelist.renumber_map[df['source']]
+            df['destination'] = input_graph.edgelist.renumber_map[df['destination']]
+
         df['overlap_coeff'] = result
 
         return df