+++ python/cugraph/tests/test_katz_centrality.py
@@ -42,9 +42,9 @@ def topKVertices(katz, col, k):
 
 
 def calc_katz(graph_file):
-    M = utils.read_csv_file(graph_file)
-    G = cugraph.Graph()
-    G.add_edge_list(M['0'], M['1'])
+    cu_M = utils.read_csv_file(graph_file)
+    G = cugraph.DiGraph()
+    G.from_cudf_edgelist(cu_M, source='0', target='1')
 
     largest_out_degree = G.degrees().nlargest(n=1, columns='out_degree')
     largest_out_degree = largest_out_degree['out_degree'][0]